refactor(llm-proxy): extract shared body helpers to drop 3x duplication
Each adapter (openai, anthropic, mistral) had its own copies of
`substituteModelJson`, `extractUsageObject`, `numberOrUndefined`, and
`parseSseDataFrame` — identical (or near-identical for the SSE one)
across all three. The 2x duplication that already existed before this
PR became 3x when Mistral landed.

Move them to `helpers.ts` and import from each adapter. The OpenAI
`[DONE]` filter in `parseSseDataFrame` is a no-op for Anthropic streams
(which never emit `[DONE]`), so the unified helper is a strict subset
of both prior behaviours — no test regressions.

Net: −48 lines across the four files, even with three concrete adapters.
Adapters now contain only the truly protocol-specific bits (which
fields to read out of `usage`, which auth header to inject, which
inbound headers to forward).
