+++ apps/api/src/services/llm-proxy/anthropic.ts
@@ -22,7 +22,12 @@
  */
 
 import type { LlmProxyAdapter, UpstreamUsage } from "./types.ts";
-import { logger } from "../../lib/logger.ts";
+import {
+  extractUsageObject,
+  numberOrUndefined,
+  parseSseDataFrame,
+  substituteModelJson,
+} from "./helpers.ts";
 
 const HEADERS_TO_FORWARD = new Set([
   "anthropic-version",
@@ -116,53 +121,4 @@ function merge(a: UpstreamUsage | null, b: UpstreamUsage): UpstreamUsage {
     cacheReadTokens: b.cacheReadTokens ?? a.cacheReadTokens,
     cacheWriteTokens: b.cacheWriteTokens ?? a.cacheWriteTokens,
   };
-}
-
-function substituteModelJson(rawBody: Uint8Array, realModelId: string): Uint8Array {
-  const text = new TextDecoder().decode(rawBody);
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(text);
-  } catch (err) {
-    logger.warn("llm-proxy: anthropic request body is not JSON — forwarding as-is", {
-      error: err instanceof Error ? err.message : String(err),
-    });
-    return rawBody;
-  }
-  if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) {
-    return rawBody;
-  }
-  (parsed as Record<string, unknown>)["model"] = realModelId;
-  return new TextEncoder().encode(JSON.stringify(parsed));
-}
-
-function extractUsageObject(body: unknown): Record<string, unknown> | null {
-  if (!body || typeof body !== "object") return null;
-  const u = (body as Record<string, unknown>)["usage"];
-  if (!u || typeof u !== "object") return null;
-  return u as Record<string, unknown>;
-}
-
-function numberOrUndefined(v: unknown): number | undefined {
-  return typeof v === "number" && Number.isFinite(v) ? v : undefined;
-}
-
-function parseSseDataFrame(chunk: string): unknown | null {
-  // Anthropic frames look like:
-  //   event: message_start
-  //   data: {"type":"message_start", …}
-  //
-  // We only need the payload from `data: …` lines.
-  const lines = chunk.split("\n");
-  const data: string[] = [];
-  for (const line of lines) {
-    if (line.startsWith("data:")) data.push(line.slice(5).trim());
-  }
-  const payload = data.join("");
-  if (!payload) return null;
-  try {
-    return JSON.parse(payload);
-  } catch {
-    return null;
-  }
 }
